리펙토링: External Channel Service가 데이터 별로 생성하도록 리펙토링
기존엔 External Service가 adapter와 commander라고 하는 클래스를 주입받아 같은 동작을 수행할 수 있도록 작성했었다. 우려 사항은 데이터에 따라서 특정 동작을 추가적으로 수행하는 상황이다. 해당 동작이 해당 데이터 전반적으로 사용되는 동작이라면 서비스 단에서 그 동작을 수행해 주는 것이 이해하기 더 좋은 동작일 것이다. 이러한 생각을 바탕으로 Commander를 삭제하고, 데이터별 서비스에서 동작들을 initialize하도록 수정했다.
