+++ app/services/external_channel/product/external_product_service.rb
@@ -1,7 +1,8 @@
 module ExternalChannel
   module Product
-    class ExternalChannelProductCommander < ExternalChannelDataCommander
-      def initialize
+    class ExternalProductService < ExternalDataService
+      def initialize(channel_adapter)
+        super(channel_adapter)
         set_saver!(ExternalChannelProductSaver.new)
         set_validator!(ExternalChannelProductValidator.new)
       end
